requirements.txt
django>=4.2,<4.3
djangorestframework
mysqlclient==2.1.1
django-cors-headers
numpy
dj-rest-auth
django-allauth
requests
reportlab
